+++ library/src/main/java/com/uploadcare/android/library/upload/UrlUploader.kt
@@ -20,7 +20,7 @@ import kotlin.math.pow
 @Suppress("unused")
 class UrlUploader(private val client: UploadcareClient, private val sourceUrl: String) : Uploader {
 
-    private var store = "auto"
+    private var store = Uploader.DEFAULT_STORE_MODE
 
     private var filename: String? = null
 
@@ -81,11 +81,16 @@ class UrlUploader(private val client: UploadcareClient, private val sourceUrl: S
     /**
      * Store the file upon uploading.
      *
-     * @param store is set true - store the file upon uploading. Requires “automatic file storing” setting to be enabled.
+     * @param store is set true - store the file upon uploading.
      * is set false - do not store file upon uploading.
+     * is set null - use auto-store setting.
      */
-    override fun store(store: Boolean): UrlUploader {
-        this.store = if (store) 1.toString() else 0.toString()
+    override fun store(store: Boolean?): UrlUploader {
+        this.store = when (store) {
+            true -> 1.toString()
+            false -> 0.toString()
+            else -> Uploader.DEFAULT_STORE_MODE
+        }
         return this
     }